put the collector/py package behind the cpython tag
