+++ pkg/collector/py/datadog_agent.c
@@ -1,3 +1,5 @@
+// +build cpython
+
 #include "datadog_agent.h"
 
 // Functions
