+++ pkg/collector/py/datadog_agent.h
@@ -1,3 +1,5 @@
+// +build cpython
+
 #ifndef DATADOG_HEADER
 #define DATADOG_HEADER
